requirements.txt
# Core data
pandas>=2.0.0
numpy>=1.24.0

# ML
scikit-learn>=1.6.0
xgboost>=2.0.0
shap>=0.44.0
imbalanced-learn>=0.12.0

# NLP
nltk>=3.8.0
textblob>=0.18.0
langdetect>=1.0.9

# Anthropic API
anthropic>=0.40.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0

# Utilities
python-dotenv>=1.0.0
jupyter>=1.0.0
pytest>=8.0.0
python-dateutil>=2.9.0
